// src/Models/Contract/ProductModel.ts
export type ProductModel = {
  barcode: string;
  product_name: string;
  price: number;
  balance: number;
  to_an_empty_warehouse: number;
  discount: number;
  balance_cost: number;
  sold_amount: number;
  last_change_price_date: string;
  finished_price: number;
  product_brand: string;
  sold_quantity: number;
  promo: number;
  image: string;
  lost_profit: number;
  nm_id: number;
  sales_per_day: number;
};